+++ spring-web/src/main/java/org/springframework/web/bind/support/WebExchangeBindException.java
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.web.bind;
+
+package org.springframework.web.bind.support;
 
 import java.beans.PropertyEditor;
 import java.util.List;
